% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndGOterm.R
\name{IndGOterm}
\alias{IndGOterm}
\title{The main function of individPath}
\usage{
IndGOterm(
  stable.pair,
  term,
  patients,
  threads = 1L,
  gene.minimum = 1L,
  threshold = 0.05,
  term_gene_list,
  non_redun
)
}
\arguments{
\item{stable.pair}{a matrix or data.frame of stable pairs from normal samples with two columns, the expression of the genes in the first column is higher than that in the second column.}

\item{term}{a matrix or data.frame contains the informations of the terms in the Gene Ontology database. Download from http://archive.geneontology.org/lite/2019-07-06/. By default the variables are taken from environment(formula).}

\item{patients}{a matrix of data.frame of tumor samples , the first column must be the geneID ,and tumor samples start with the second column.}

\item{threads}{an integer value to make sure how many threads you will use to complete the computation}

\item{gene.minimum}{an integer value to control the minimum amount of the genes of the terms. The deault is 1.}

\item{threshold}{a numeric value which is used to control false discovery rate under the p_value of the chip-square test or fisher exact probability test , default is 0.05.}

\item{term_gene_list}{a list of term genes. list type. each list is a vector type, containing all the genes of a term.By default the variables are "BPtermgene_list" and taken from environment(formula).}

\item{non_redun}{a vector of termID used to the screening of the personalized dysregulation terms.}
}
\value{
a list contained two data.frame , individual dyregulated pathway with term ID and individual dyregulated pathway with term name
}
\description{
The function is used to screen personalized dysregulation terms.
this function support parallel computation . You need to set thread numbers to make sure how many threads do you need to use.
}
